Wael

+++ app/Livewire/Pcash/TillForm.php
@@ -146,7 +146,7 @@ class TillForm extends Component
             $tillAmountsIds[] = $tillAmount->id;
         }
 
-        TillAmount::whereNotIn('id', $tillAmountsIds)->delete();
+        TillAmount::whereNotIn('id', $tillAmountsIds)->where('user_id', $this->user_id)->delete();
 
         session()->flash('success', 'till has been updated successfully!');
